Update combine free block logic
In myfree, when there are 2 adjacent free blocks, the curr ptr would become curr + prev.size + sizeof(metablock). But, that means that when searching for a free block, we would still see the prev block AND the curr block.

So, changed it so that prev->size += curr->size + sizeof(metablock), effectively making the prev absorb the curr free block.

Same reason why it is now mergeNext(prev).